feat(tensor_graph): add scheduler for graph

+++ python/tvm/tensor_graph/testing/models/shufflenet.py
@@ -9,10 +9,15 @@ from tvm.tensor_graph.nn import CELoss, SGD
 import math
 import tvm
 
-def conv_bn(in_channel, out_channel, kernel_size, strides=1, padding=0, dilation=1, groups=1, use_bias=False):
+def conv_bn(in_channel, out_channel, kernel_size,
+        strides=1, padding=0, dilation=1, groups=1, use_bias=False,
+        dtype="float32", out_dtype="float32"):
     return Sequential(
-        Conv2d(in_channel, out_channel, kernel_size, stride=strides, padding=padding, dilation=dilation, groups=groups, bias=use_bias),
-        BatchNorm2d(out_channel),
+        Conv2d(in_channel, out_channel, kernel_size,
+            stride=strides, padding=padding,
+            dilation=dilation, groups=groups, bias=use_bias,
+            dtype=dtype, out_dtype=out_dtype),
+        BatchNorm2d(out_channel, dtype=dtype, out_dtype=out_dtype),
         ReLU()
     )
 
@@ -59,23 +64,30 @@ def cat_dim1(A, B):
 
 class ShuffleNetUnitA(Layer):
     """ShuffleNet unit for stride=1"""
-    def __init__(self, in_channels, out_channels, groups=3):
+    def __init__(self, in_channels, out_channels, groups=3,
+        dtype="float32", out_dtype="float32"):
         super(ShuffleNetUnitA, self).__init__()
         assert in_channels == out_channels
         assert out_channels % 4 == 0
         bottleneck_channels = out_channels // 4
         self.groups = groups
         self.group_conv1 = Conv2d(in_channels, bottleneck_channels,
-                                        1, groups=groups, stride=1)
-        self.bn2 = BatchNorm2d(bottleneck_channels)
+                                  1, groups=groups, stride=1,
+                                  dtype=dtype, out_dtype=out_dtype)
+        self.bn2 = BatchNorm2d(
+            bottleneck_channels, dtype=dtype, out_dtype=out_dtype)
         self.depthwise_conv3 = Conv2d(bottleneck_channels,
                                          bottleneck_channels,
                                          3, padding=1, stride=1,
-                                         groups=bottleneck_channels)
-        self.bn4 = BatchNorm2d(bottleneck_channels)
+                                         groups=bottleneck_channels,
+                                         dtype=dtype, out_dtype=out_dtype)
+        self.bn4 = BatchNorm2d(
+            bottleneck_channels, dtype=dtype, out_dtype=out_dtype)
         self.group_conv5 = Conv2d(bottleneck_channels, out_channels,
-                                     1, stride=1, groups=groups)
-        self.bn6 = BatchNorm2d(out_channels)
+                                     1, stride=1, groups=groups,
+                                     dtype=dtype, out_dtype=out_dtype)
+        self.bn6 = BatchNorm2d(
+            out_channels, dtype=dtype, out_dtype=out_dtype)
 
     def forward(self, x):
         out = self.group_conv1(x)
@@ -90,23 +102,30 @@ class ShuffleNetUnitA(Layer):
 
 class ShuffleNetUnitB(Layer):
     """ShuffleNet unit for stride=2"""
-    def __init__(self, in_channels, out_channels, groups=3):
+    def __init__(self, in_channels, out_channels, groups=3,
+        dtype="float32", out_dtype="float32"):
         super(ShuffleNetUnitB, self).__init__()
         out_channels -= in_channels
         assert out_channels % 4 == 0
         bottleneck_channels = out_channels // 4
         self.groups = groups
         self.group_conv1 = Conv2d(in_channels, bottleneck_channels,
-                                     1, groups=groups, stride=1)
-        self.bn2 = BatchNorm2d(bottleneck_channels)
+                                     1, groups=groups, stride=1,
+                                     dtype=dtype, out_dtype=out_dtype)
+        self.bn2 = BatchNorm2d(
+            bottleneck_channels, dtype=dtype, out_dtype=out_dtype)
         self.depthwise_conv3 = Conv2d(bottleneck_channels,
                                          bottleneck_channels,
                                          3, padding=1, stride=2,
-                                         groups=bottleneck_channels)
-        self.bn4 = BatchNorm2d(bottleneck_channels)
+                                         groups=bottleneck_channels,
+                                         dtype=dtype, out_dtype=out_dtype)
+        self.bn4 = BatchNorm2d(
+            bottleneck_channels, dtype=dtype, out_dtype=out_dtype)
         self.group_conv5 = Conv2d(bottleneck_channels, out_channels,
-                                     1, stride=1, groups=groups)
-        self.bn6 = BatchNorm2d(out_channels)
+                                     1, stride=1, groups=groups,
+                                     dtype=dtype, out_dtype=out_dtype)
+        self.bn6 = BatchNorm2d(
+            out_channels, dtype=dtype, out_dtype=out_dtype)
 
     def forward(self, x):
         out = self.group_conv1(x)
@@ -122,20 +141,32 @@ class ShuffleNetUnitB(Layer):
 
 class ShuffleNet(Layer):
     """ShuffleNet for groups=3"""
-    def __init__(self, groups=3, in_channels=3, num_classes=1000):
+    def __init__(self, groups=3, in_channels=3, num_classes=1000,
+        dtype="float32", out_dtype="float32"):
         super(ShuffleNet, self).__init__()
 
-        self.conv1 = Conv2d(in_channels, 24, 3, stride=2, padding=1)
-        stage2_seq = [ShuffleNetUnitB(24, 240, groups=3)] + \
-            [ShuffleNetUnitA(240, 240, groups=3) for i in range(3)]
+        self.conv1 = Conv2d(
+            in_channels, 24, 3, stride=2, padding=1,
+            dtype=dtype, out_dtype=out_dtype)
+        stage2_seq = [
+            ShuffleNetUnitB(
+                24, 240, groups=3, dtype=dtype, out_dtype=out_dtype)] + \
+            [ShuffleNetUnitA(
+                240, 240, groups=3, dtype=dtype, out_dtype=out_dtype) for i in range(3)]
         self.stage2 = Sequential(*stage2_seq)
-        stage3_seq = [ShuffleNetUnitB(240, 480, groups=3)] + \
-            [ShuffleNetUnitA(480, 480, groups=3) for i in range(7)]
+        stage3_seq = [
+            ShuffleNetUnitB(
+                240, 480, groups=3, dtype=dtype, out_dtype=out_dtype)] + \
+            [ShuffleNetUnitA(
+                480, 480, groups=3, dtype=dtype, out_dtype=out_dtype) for i in range(7)]
         self.stage3 = Sequential(*stage3_seq)
-        stage4_seq = [ShuffleNetUnitB(480, 960, groups=3)] + \
-                     [ShuffleNetUnitA(960, 960, groups=3) for i in range(3)]
+        stage4_seq = [
+            ShuffleNetUnitB(
+                480, 960, groups=3, dtype=dtype, out_dtype=out_dtype)] + \
+                     [ShuffleNetUnitA(
+                         960, 960, groups=3, dtype=dtype, out_dtype=out_dtype) for i in range(3)]
         self.stage4 = Sequential(*stage4_seq)
-        self.fc = Linear(960, num_classes)
+        self.fc = Linear(960, num_classes, dtype=dtype, out_dtype=out_dtype)
 
     def forward(self, x):
         net = self.conv1(x)
